Movy/Backend/requirements.txt
Django==4.2
djangorestframework==3.14
django-cors-headers==4.1.0
jazzmin==2.6.0
djangorestframework-simplejwt>=5.2.2
redis>=3.5.3